requirements.txt

Django==3.2.10
django-recaptcha3
Pillow
requests
sorl-thumbnail
lxml
django-ckeditor
django-cleanup
django-allauth
whitenoise
yookassa
telepot
pandas
pytils
beautifulsoup4
fake-useragent
zeep